requirements.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pyyaml==6.0.1
kubernetes==28.1.0
markdown==3.5.1
jsonschema==4.20.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1